chatbot/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
dialogflow==0.8.0
enum34==1.1.6
google-api-core==1.16.0
google-auth==1.13.1
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.8
numpy==1.18.2
opencv-contrib-python==4.2.0.34
peewee==3.13.2
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
urllib3==1.25.8
vk-api==11.4.0
